requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
transformers>=4.11.0
python-dotenv>=0.19.0

# Desarrollo
pytest>=6.2.5
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.4.0
isort>=5.12.0
pre-commit>=3.3.0
httpx>=0.24.0
uvicorn>=0.15.0
fastapi>=0.68.0
uvicorn>=0.23.0
fastapi>=0.100.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
